Read-er supporting multiple input encodings and xml input preproc
Note that this is *not* intended to be used as an xml parser,
but rather to do some preprocessing of the input data that's
supposed to occur *before* data reaches the xml parser.

The most notable of these are eliminating BOM bytes and replacing
newlines.

Note that newline replacement has not yet been implemented.  And all the
test fail.  But transcoding input encodings to UTF-8 works!  We're just
losing the first two chars or so of the input.